requirements.txt
black==24.8.0
fastapi==0.115.0
flake8==7.1.1
openai==1.51.0
pre-commit==3.8.0
python-dotenv==1.0.1
uvicorn==0.31.0
